engine: the provider interface nobody outside could implement
provider.Database.ConnString returned secrets.Value from
engine/internal/secrets. An implementation has to name the return type, and
there is no spelling of that type an outside package is allowed to use: naming
it without the import is "undefined: secrets", and adding the import is "use of
internal package ... not allowed". The interface the 1.0.0 notes call an
integration surface, meant to be written outside this repository, could not be
written outside this repository at all.

It compiled here, it reviewed as correct, and it would have failed on the first
line of the first real provider anybody wrote. The tag would have held it there
for the whole of version 1.

The type moves to engine/pkg/secret, which is public for exactly this reason,
and engine/internal/secrets.Value is an alias for it. An alias rather than a
wrapper: the two are the same type to the compiler, so every call site, struct
field and type assertion inside the module keeps working and there is no
conversion to forget. GoldenSpec, EnvSpec, ServiceSpec and Config carried the
same type in exported fields and move with it.
