+++ engine/internal/secrets/value.go
@@ -1,125 +1,31 @@
-// Package secrets holds the engine's secret material and the type that makes
-// leaking it hard.
+// Package secrets holds the engine's secret material: where a credential is
+// loaded from, how it is resolved, and the keyring it can be kept in.
 //
-// A Value carries a secret string but renders as "[redacted]" through every
-// path the standard library uses to turn a value into text: fmt verbs, the
-// Stringer interface, JSON and YAML marshalling, and slog attributes. Reading
-// the plaintext requires calling Reveal, which is easy to grep for and easy to
-// review. The rest of the engine passes Value around freely without any call
-// site needing to remember to redact.
+// The type that carries the credential itself lives in engine/pkg/secret, one
+// directory outward and outside internal, and this file is the alias that lets
+// the rest of the engine keep spelling it secrets.Value. It is public there
+// because engine/pkg/provider names it in the interfaces a provider has to
+// implement, and a type an outside package cannot name is an interface an
+// outside package cannot implement. The reasoning is written out where the
+// type is.
+//
+// The alias is an alias and not a wrapper on purpose: secrets.Value and
+// secret.Value are the same type to the compiler, so every existing call site,
+// struct field and type assertion keeps working and no conversion exists to
+// forget.
 package secrets
 
-import (
-	"crypto/sha256"
-	"crypto/subtle"
-	"encoding/hex"
-	"fmt"
-	"io"
-	"log/slog"
-)
+import "github.com/antifailure/antifailure/engine/pkg/secret"
+
+// Value is a secret string that does not print itself. See engine/pkg/secret.
+type Value = secret.Value
 
 // Redacted is what a Value renders as anywhere text is produced.
-const Redacted = "[redacted]"
-
-// Value is a secret string that does not print itself.
-//
-// The zero Value is empty and safe to use. Value is comparable only through
-// Equal, which is constant time.
-type Value struct {
-	// v is unexported and never embedded in a struct that gets marshalled
-	// directly, so the only way to observe it is Reveal.
-	v string
-	// source records where the value came from, for audit events. It names a
-	// source such as "keyring" or "env" and never contains the secret.
-	source string
-}
+const Redacted = secret.Redacted
 
 // New returns a Value holding s.
-func New(s string) Value { return Value{v: s} }
+func New(s string) Value { return secret.New(s) }
 
 // NewFrom returns a Value holding s, tagged with the source that produced it.
 // The source appears in audit events; the value never does.
-func NewFrom(s, source string) Value { return Value{v: s, source: source} }
-
-// Reveal returns the plaintext.
-//
-// Every call site is a place a secret can escape. Keep them few, keep them
-// close to the boundary that needs the plaintext (a provider client, a
-// subprocess environment, a connection string), and never pass the result on.
-func (v Value) Reveal() string { return v.v }
-
-// Source reports where the value was loaded from, or the empty string when it
-// was not tagged.
-func (v Value) Source() string { return v.source }
-
-// IsZero reports whether the Value holds no secret.
-func (v Value) IsZero() bool { return v.v == "" }
-
-// Len reports the length of the secret in bytes. It is safe to log: a length
-// is not a secret and it helps diagnose a truncated credential.
-func (v Value) Len() int { return len(v.v) }
-
-// Equal reports whether two values hold the same secret, in constant time so
-// that comparing does not leak the contents through timing.
-func (v Value) Equal(other Value) bool {
-	return subtle.ConstantTimeCompare([]byte(v.v), []byte(other.v)) == 1
-}
-
-// Fingerprint returns the first eight hex characters of the SHA-256 of the
-// secret, or the empty string for a zero Value.
-//
-// It exists so that an operator can tell whether the credential in Key Vault
-// changed without ever printing it. Eight hex characters is thirty-two bits:
-// enough to notice a rotation, far too little to attack the preimage.
-func (v Value) Fingerprint() string {
-	if v.v == "" {
-		return ""
-	}
-	sum := sha256.Sum256([]byte(v.v))
-	return hex.EncodeToString(sum[:])[:8]
-}
-
-// String renders the redaction marker.
-func (v Value) String() string { return Redacted }
-
-// GoString renders the redaction marker, so %#v is safe too.
-func (v Value) GoString() string { return Redacted }
-
-// Format renders the redaction marker for every fmt verb.
-//
-// Implementing fmt.Formatter rather than only fmt.Stringer matters: with only
-// a Stringer, %x hex encodes the marker and %d prints a badly formatted verb
-// notice. Neither leaks, but neither is readable either. Taking over Format
-// means the marker appears verbatim no matter what verb a call site reaches
-// for, including a wrong one.
-func (v Value) Format(f fmt.State, verb rune) {
-	if verb == 'q' {
-		_, _ = io.WriteString(f, `"`+Redacted+`"`)
-		return
-	}
-	_, _ = io.WriteString(f, Redacted)
-}
-
-// MarshalJSON renders the redaction marker as a JSON string.
-func (v Value) MarshalJSON() ([]byte, error) {
-	return []byte(`"` + Redacted + `"`), nil
-}
-
-// UnmarshalJSON reads a secret from JSON. Round tripping a marshalled Value
-// therefore yields the marker rather than the secret, which is intended: a
-// secret must be loaded from a secret source, never from a document that has
-// already been through a marshaller.
-func (v *Value) UnmarshalJSON(b []byte) error {
-	s := string(b)
-	if len(s) >= 2 && s[0] == '"' && s[len(s)-1] == '"' {
-		s = s[1 : len(s)-1]
-	}
-	v.v = s
-	return nil
-}
-
-// MarshalYAML renders the redaction marker.
-func (v Value) MarshalYAML() (any, error) { return Redacted, nil }
-
-// LogValue renders the redaction marker for log/slog.
-func (v Value) LogValue() slog.Value { return slog.StringValue(Redacted) }
+func NewFrom(s, source string) Value { return secret.NewFrom(s, source) }
